Extracted point components from site files
Created basic new site wizard component
