+++ src/app/components/sites/pages/delete/site.component.ts
@@ -1,13 +1,7 @@
 import { Component, OnInit } from "@angular/core";
 import { ActivatedRoute, Router } from "@angular/router";
 import { projectResolvers } from "@baw-api/project/projects.service";
-import { regionResolvers } from "@baw-api/region/regions.service";
 import { siteResolvers, SitesService } from "@baw-api/site/sites.service";
-import {
-  deletePointMenuItem,
-  pointMenuItem,
-  pointsCategory,
-} from "@components/sites/points.menus";
 import {
   deleteSiteMenuItem,
   siteMenuItem,
@@ -18,16 +12,12 @@ import {
   FormTemplate,
 } from "@helpers/formTemplate/formTemplate";
 import { Project } from "@models/Project";
-import { Region } from "@models/Region";
 import { Site } from "@models/Site";
 import { PermissionsShieldComponent } from "@shared/permissions-shield/permissions-shield.component";
 import { WidgetMenuItem } from "@shared/widget/widgetItem";
 import { List } from "immutable";
 import { ToastrService } from "ngx-toastr";
-import {
-  pointMenuItemActions,
-  siteMenuItemActions,
-} from "../details/details.component";
+import { siteMenuItemActions } from "../details/site.component";
 
 const projectKey = "project";
 const regionKey = "region";
@@ -72,20 +62,6 @@ class SiteDeleteComponent extends FormTemplate<Site> implements OnInit {
   }
 }
 
-@Component({
-  selector: "app-points-delete",
-  templateUrl: "./delete.component.html",
-})
-class PointDeleteComponent extends SiteDeleteComponent {
-  public get region(): Region {
-    return this.models.region as Region;
-  }
-
-  protected redirectionPath() {
-    return this.region.viewUrl;
-  }
-}
-
 SiteDeleteComponent.LinkComponentToPageInfo({
   category: sitesCategory,
   menus: {
@@ -98,17 +74,4 @@ SiteDeleteComponent.LinkComponentToPageInfo({
   },
 }).AndMenuRoute(deleteSiteMenuItem);
 
-PointDeleteComponent.LinkComponentToPageInfo({
-  category: pointsCategory,
-  menus: {
-    actions: List([pointMenuItem, ...pointMenuItemActions]),
-    actionsWidget: new WidgetMenuItem(PermissionsShieldComponent, {}),
-  },
-  resolvers: {
-    [projectKey]: projectResolvers.show,
-    [regionKey]: regionResolvers.show,
-    [siteKey]: siteResolvers.show,
-  },
-}).AndMenuRoute(deletePointMenuItem);
-
-export { SiteDeleteComponent, PointDeleteComponent };
+export { SiteDeleteComponent };
