+++ src/app/components/sites/points.module.ts
@@ -3,10 +3,10 @@ import { RouterModule } from "@angular/router";
 import { SitesModule } from "@components/sites/sites.module";
 import { GetRouteConfigForPage } from "@helpers/page/pageRouting";
 import { SharedModule } from "@shared/shared.module";
-import { PointDeleteComponent } from "./pages/delete/delete.component";
-import { PointDetailsComponent } from "./pages/details/details.component";
-import { PointEditComponent } from "./pages/edit/edit.component";
-import { PointNewComponent } from "./pages/new/new.component";
+import { PointDeleteComponent } from "./pages/delete/point.component";
+import { PointDetailsComponent } from "./pages/details/point.component";
+import { PointEditComponent } from "./pages/edit/point.component";
+import { PointNewComponent } from "./pages/new/point.component";
 import { pointsRoute } from "./points.menus";
 
 const components = [
