+++ src/app/components/sites/sites.module.ts
@@ -3,11 +3,12 @@ import { RouterModule } from "@angular/router";
 import { GetRouteConfigForPage } from "@helpers/page/pageRouting";
 import { MapModule } from "@shared/map/map.module";
 import { SharedModule } from "@shared/shared.module";
-import { SiteDeleteComponent } from "./pages/delete/delete.component";
-import { SiteDetailsComponent } from "./pages/details/details.component";
-import { SiteEditComponent } from "./pages/edit/edit.component";
+import { SiteDeleteComponent } from "./pages/delete/site.component";
+import { SiteDetailsComponent } from "./pages/details/site.component";
+import { SiteEditComponent } from "./pages/edit/site.component";
 import { HarvestComponent } from "./pages/harvest/harvest.component";
-import { SiteNewComponent } from "./pages/new/new.component";
+import { SiteNewComponent } from "./pages/new/site.component";
+import { WizardComponent } from "./pages/wizard/wizard.component";
 import { SiteComponent } from "./site/site.component";
 import { sitesRoute } from "./sites.menus";
 
@@ -18,6 +19,7 @@ const components = [
   SiteDetailsComponent,
   SiteEditComponent,
   SiteNewComponent,
+  WizardComponent,
 ];
 
 const routes = sitesRoute.compileRoutes(GetRouteConfigForPage);
